% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandCircleDistance.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Calculate the distance between two points}
\usage{
calculateDistance(latitude1, latitude2, longitude1, longitude2)
}
\arguments{
\item{longitude2}{}
}
\value{

}
\description{
Calculate thet distance between two points, specified by their respective latitude and longitude.
This function uses the grand circle distance to calculate the distance accounting for the curvature of the earth.
The function can be inaccurate at distance <2km, but this should not affect use with the weather model, since the resolution is not that precise.
}
