% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeather.R
\name{getWeather}
\alias{getWeather}
\title{Get the weather}
\usage{
getWeather(
  longitude,
  latitude,
  timefrom,
  timeto,
  unit = "hour",
  variables = NA
)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Get the weather at a given site (or the closest downloaded site), in the specified time period.
The Copernicus weather data is in hours, but may be aggregated into days.
}
\details{
Variables that have been downloaded from the ERA5 dataset include:
  d2m - 2-metre dewpoint temperature (K)
  t2m - 2-metre temperature (K)
  sp - surface pressure (Pa)
  tp - total precipitation (m)
  ssrd - surface solar radiation downwards (J/m^2)

Variables that are extrapolated are:
  RH - relative humidity
}
